feat(ui): Email Attachments
This is a rough walkthrough of thinking about email
attachments in a "modular" fashion.

Basic idea is that Maxwell or/and Modular will look at the
type of email attachment and select & compose the
best module to preview that given attachment.

I've created a basic Attachment Data Model where
the "type" of attachment is specified, and a example
Youtube Preview "module" (which is still just a widget
for now) to illustrate the workflow.

Furthermore, I've added a "fallback"
module that renders attachments that might have
no corresponding module with a simple icon
(much like Inbox/Gmail).

The ThreadListItem will provide a "slot" for every
attachment. Which should be filled by appropriate
modules based on the type of attachment. This is
still a TODO.

This is more exploratory so I would like your
feedback.

Change-Id: Ie0be4f8e1112afa439925a3494bf8c518ad768f9